refactoring: #12 메서드 명 변경 - rentalBook -> createRentalBook
