+++ src/main/java/flab/library/rental/service/command/RentalService.java
@@ -1,4 +1,4 @@
-package flab.library.rental.service;
+package flab.library.rental.service.command;
 
 import flab.library.book.domain.entity.Book;
 import flab.library.rental.domain.entity.Rental;
@@ -6,5 +6,5 @@ import flab.library.user.domain.entity.LibUser;
 import java.time.LocalDateTime;
 
 public interface RentalService {
-  Rental rentalBook(LibUser user, Book book, LocalDateTime endDate);
+  Rental createRentalBook(LibUser user, Book book, LocalDateTime endDate);
 }
